+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Change categories:
 9. pdr: Remove pldm_pdr_add_fru_record_set()
 10. pdr: Remove pldm_entity_association_pdr_add_from_node()
 11. pdr: Make find_entity_ref_in_tree() static
+12. pdr: Make entity_association_tree_find() static
 
 ### Fixed